refactor: 패키지 구조 구체화
