+++ backend/src/main/java/reviewme/keyword/KeywordController.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
+import reviewme.keyword.dto.response.KeywordsResponse;
 
 @RestController
 @RequiredArgsConstructor
