+++ backend/src/main/java/reviewme/keyword/dto/response/KeywordResponse.java
@@ -1,4 +1,4 @@
-package reviewme.keyword;
+package reviewme.keyword.dto.response;
 
 public record KeywordResponse(
         long id,
