+++ backend/src/main/java/reviewme/keyword/dto/response/KeywordsResponse.java
@@ -1,4 +1,4 @@
-package reviewme.keyword;
+package reviewme.keyword.dto.response;
 
 import java.util.List;
