+++ backend/src/main/java/reviewme/member/ReviewerGroupController.java
@@ -5,6 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
+import reviewme.member.dto.response.ReviewerGroupResponse;
 
 @RestController
 @RequiredArgsConstructor
