+++ backend/src/main/java/reviewme/member/dto/response/ReviewerGroupResponse.java
@@ -1,4 +1,4 @@
-package reviewme.member;
+package reviewme.member.dto.response;
 
 import java.time.LocalDateTime;
