--- backend/src/main/java/reviewme/review/ContentResponse.java
@@ -1,8 +1,0 @@
-package reviewme.review;
-
-public record ContentResponse(
-        long id,
-        String question,
-        String answer
-) {
-}
