+++ backend/src/main/java/reviewme/review/ReviewController.java
@@ -8,6 +8,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import reviewme.review.dto.request.CreateReviewRequest;
+import reviewme.review.dto.response.ReviewResponse;
 
 @RestController
 @RequiredArgsConstructor
