--- backend/src/main/java/reviewme/review/ReviewResponse.java
@@ -1,16 +1,0 @@
-package reviewme.review;
-
-import java.util.List;
-import reviewme.keyword.KeywordResponse;
-import reviewme.member.MemberResponse;
-import reviewme.member.ReviewerGroupResponse;
-
-public record ReviewResponse(
-        long id,
-//        LocalDateTime createdAt,
-        MemberResponse reviewer,
-        ReviewerGroupResponse reviewerGroup,
-        List<ContentResponse> contents,
-        List<KeywordResponse> keywords
-) {
-}
