+++ backend/src/main/java/reviewme/review/dto/response/ReviewContentResponse.java
@@ -1,0 +1,8 @@
+package reviewme.review.dto.response;
+
+public record ReviewContentResponse(
+        long id,
+        String question,
+        String answer
+) {
+}
