+++ backend/src/test/java/reviewme/review/ReviewServiceTest.java
@@ -12,6 +12,9 @@ import reviewme.member.Member;
 import reviewme.member.MemberRepository;
 import reviewme.member.ReviewerGroup;
 import reviewme.member.ReviewerGroupRepository;
+import reviewme.review.dto.request.CreateReviewContentRequest;
+import reviewme.review.dto.request.CreateReviewRequest;
+import reviewme.review.dto.response.ReviewResponse;
 import reviewme.support.ServiceTest;
 
 @ServiceTest
